Basic map functions are working properly
Pins are dropped to all houses in a 0.25 mile radius
of your location. The pin for your location is light blue.
Visited pins change color from green to dark blue. If you
click on the info window for the pin, a pop up of the house info
will appear. Default zoom is disabled, only can use zoom buttons provided
to zoom in. Function getHouseInfoList returns list of addresses of
surrounding houses.